+++ src/file-manager/commands/cp.js
@@ -1,7 +1,7 @@
 import { createReadStream, createWriteStream } from 'node:fs';
 import * as path from 'node:path';
-import { doesFileExist } from '../../../utils/doesFileExist.js';
-import { ERROR_MESSAGES } from '../../../consts.js';
+import { doesFileExist } from '../../utils/doesFileExist.js';
+import { ERROR_MESSAGES } from '../../consts.js';
 
 export default async (args) => {
     if (!args.length) {
