+++ src/file-manager/commands/ls.js
