+++ src/file-manager/commands/up.js
